backend/my_diplom/backend/serializers.py
from rest_framework import serializers

from my_diplom.backend.models import Shop, Product


class ShopSerializer(serializers.ModelSerializer):
    class Meta:
        model = Shop
        fields = ('id', 'name', 'state',)


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ('id', 'name', 'category', 'quantity', 'price', 'shop',)
