backend/requirements.txt
django
setuptools
djangorestframework
psycopg2-binary